Using generic format specifier for logging pointers
This change uses generic format specifier for logging the
pointers which can be 64 bit or 32 bit depending upon the
system used for compilation.
Also address of operator is used before casting a int variable
to void pointer.

+++ osi/include/data_dispatcher.h
@@ -26,7 +26,7 @@
 #define DISPATCHER_NAME_MAX 16
 
 typedef struct data_dispatcher_t data_dispatcher_t;
-typedef uint32_t data_dispatcher_type_t;
+typedef uintptr_t data_dispatcher_type_t;
 
 // Creates a new data dispatcher object, with the given name.
 // The returned object must be freed by calling |data_dispatcher_free|.
